Cardano.Crypto.DSIGN.Class

Description

Abstract digital signatures.

DSIGN algorithm class

class (Typeable v, Show (VerKeyDSIGN v), Eq (VerKeyDSIGN v), Show (SignKeyDSIGN v), Show (SigDSIGN v), Eq (SigDSIGN v), NoThunks (SigDSIGN v), NoThunks (SignKeyDSIGN v), NoThunks (VerKeyDSIGN v), KnownNat (SeedSizeDSIGN v), KnownNat (SizeVerKeyDSIGN v), KnownNat (SizeSignKeyDSIGN v), KnownNat (SizeSigDSIGN v)) ⇒ DSIGNAlgorithm v where

The pure DSIGN API, which supports the full set of DSIGN operations, but does not allow for secure forgetting of private keys.

Minimal complete definition

algorithmNameDSIGN, deriveVerKeyDSIGN, signDSIGN, verifyDSIGN, genKeyDSIGN, rawSerialiseVerKeyDSIGN, rawSerialiseSignKeyDSIGN, rawSerialiseSigDSIGN, rawDeserialiseVerKeyDSIGN, rawDeserialiseSignKeyDSIGN, rawDeserialiseSigDSIGN

Associated Types

type SeedSizeDSIGN v ∷ Nat

type SizeVerKeyDSIGN v ∷ Nat

type SizeSignKeyDSIGN v ∷ Nat

type SizeSigDSIGN v ∷ Nat

data VerKeyDSIGN v ∷ Type

data SignKeyDSIGN v ∷ Type

data SigDSIGN v ∷ Type

type ContextDSIGN v ∷ Type

Context required to run the DSIGN algorithm

Unit by default (no context required)

type ContextDSIGN v = ()

type Signable v ∷ Type → Constraint

type Signable v = Empty

Methods

algorithmNameDSIGN ∷ proxy v → String

deriveVerKeyDSIGN ∷ SignKeyDSIGN v → VerKeyDSIGN v

hashVerKeyDSIGN ∷ HashAlgorithm h ⇒ VerKeyDSIGN v → Hash h (VerKeyDSIGN v)

signDSIGN ∷ (Signable v a, HasCallStack) ⇒ ContextDSIGN v → a → SignKeyDSIGN v → SigDSIGN v

verifyDSIGN ∷ (Signable v a, HasCallStack) ⇒ ContextDSIGN v → VerKeyDSIGN v → a → SigDSIGN v → Either String ()

genKeyDSIGN ∷ Seed → SignKeyDSIGN v

Note that this function may error (with SeedBytesExhausted) if the provided seed is not long enough. Callers should ensure that the seed has is at least seedSizeDSIGN bytes long.

rawSerialiseVerKeyDSIGN ∷ VerKeyDSIGN v → ByteString

rawSerialiseSignKeyDSIGN ∷ SignKeyDSIGN v → ByteString

rawSerialiseSigDSIGN ∷ SigDSIGN v → ByteString

rawDeserialiseVerKeyDSIGN ∷ ByteString → Maybe (VerKeyDSIGN v)

rawDeserialiseSignKeyDSIGN ∷ ByteString → Maybe (SignKeyDSIGN v)

rawDeserialiseSigDSIGN ∷ ByteString → Maybe (SigDSIGN v)

data Seed

A seed contains a finite number of bytes, and is used for seeding cryptographic algorithms including key generation.

This is not itself a PRNG, but can be used to seed a PRNG.

seedSizeDSIGN ∷ ∀ v proxy. DSIGNAlgorithm v ⇒ proxy v → Word

The upper bound on the Seed size needed by genKeyDSIGN

sizeVerKeyDSIGN ∷ ∀ v proxy. DSIGNAlgorithm v ⇒ proxy v → Word

sizeSignKeyDSIGN ∷ ∀ v proxy. DSIGNAlgorithm v ⇒ proxy v → Word

sizeSigDSIGN ∷ ∀ v proxy. DSIGNAlgorithm v ⇒ proxy v → Word

MLocked DSIGN algorithm class

class (DSIGNAlgorithm v, NoThunks (SignKeyDSIGNM v)) ⇒ DSIGNMAlgorithm v where

Associated Types

data SignKeyDSIGNM v ∷ Type

Methods

deriveVerKeyDSIGNM ∷ (MonadThrow m, MonadST m) ⇒ SignKeyDSIGNM v → m (VerKeyDSIGN v)

signDSIGNM ∷ (Signable v a, MonadST m, MonadThrow m) ⇒ ContextDSIGN v → a → SignKeyDSIGNM v → m (SigDSIGN v)

genKeyDSIGNMWith ∷ (MonadST m, MonadThrow m) ⇒ MLockedAllocator m → MLockedSeed (SeedSizeDSIGN v) → m (SignKeyDSIGNM v)

cloneKeyDSIGNMWith ∷ MonadST m ⇒ MLockedAllocator m → SignKeyDSIGNM v → m (SignKeyDSIGNM v)

getSeedDSIGNMWith ∷ (MonadST m, MonadThrow m) ⇒ MLockedAllocator m → Proxy v → SignKeyDSIGNM v → m (MLockedSeed (SeedSizeDSIGN v))

forgetSignKeyDSIGNMWith ∷ (MonadST m, MonadThrow m) ⇒ MLockedAllocator m → SignKeyDSIGNM v → m ()

genKeyDSIGNM ∷ (DSIGNMAlgorithm v, MonadST m, MonadThrow m) ⇒ MLockedSeed (SeedSizeDSIGN v) → m (SignKeyDSIGNM v)

cloneKeyDSIGNM ∷ (DSIGNMAlgorithm v, MonadST m) ⇒ SignKeyDSIGNM v → m (SignKeyDSIGNM v)

getSeedDSIGNM ∷ (DSIGNMAlgorithm v, MonadST m, MonadThrow m) ⇒ Proxy v → SignKeyDSIGNM v → m (MLockedSeed (SeedSizeDSIGN v))

forgetSignKeyDSIGNM ∷ (DSIGNMAlgorithm v, MonadST m, MonadThrow m) ⇒ SignKeyDSIGNM v → m ()

SignedDSIGN wrapper

newtype SignedDSIGN v a

Constructors

SignedDSIGN (SigDSIGN v)

signedDSIGN ∷ (DSIGNAlgorithm v, Signable v a) ⇒ ContextDSIGN v → a → SignKeyDSIGN v → SignedDSIGN v a

verifySignedDSIGN ∷ (DSIGNAlgorithm v, Signable v a, HasCallStack) ⇒ ContextDSIGN v → VerKeyDSIGN v → a → SignedDSIGN v a → Either String ()

CBOR encoding and decoding

encodeVerKeyDSIGN ∷ DSIGNAlgorithm v ⇒ VerKeyDSIGN v → Encoding

decodeVerKeyDSIGN ∷ ∀ v s. DSIGNAlgorithm v ⇒ Decoder s (VerKeyDSIGN v)

encodeSignKeyDSIGN ∷ DSIGNAlgorithm v ⇒ SignKeyDSIGN v → Encoding

decodeSignKeyDSIGN ∷ ∀ v s. DSIGNAlgorithm v ⇒ Decoder s (SignKeyDSIGN v)

encodeSigDSIGN ∷ DSIGNAlgorithm v ⇒ SigDSIGN v → Encoding

decodeSigDSIGN ∷ ∀ v s. DSIGNAlgorithm v ⇒ Decoder s (SigDSIGN v)

encodeSignedDSIGN ∷ DSIGNAlgorithm v ⇒ SignedDSIGN v a → Encoding

decodeSignedDSIGN ∷ DSIGNAlgorithm v ⇒ Decoder s (SignedDSIGN v a)

Encoded Size expresssions

encodedVerKeyDSIGNSizeExpr ∷ ∀ v. DSIGNAlgorithm v ⇒ Proxy (VerKeyDSIGN v) → Size

Size expression for VerKeyDSIGN which is using sizeVerKeyDSIGN encoded as Size.

encodedSignKeyDSIGNSizeExpr ∷ ∀ v. DSIGNAlgorithm v ⇒ Proxy (SignKeyDSIGN v) → Size

Size expression for SignKeyDSIGN which is using sizeSignKeyDSIGN encoded as Size.

encodedSigDSIGNSizeExpr ∷ ∀ v. DSIGNAlgorithm v ⇒ Proxy (SigDSIGN v) → Size

Size expression for SigDSIGN which is using sizeSigDSIGN encoded as Size.

Helper

failSizeCheck ∷ MonadFail m ⇒ String → String → ByteString → Word → m a

Helper function that always fails, but it provides a different message whenever expected size does not match.

Unsound CBOR encoding and decoding of MLocked DSIGN keys

class DSIGNMAlgorithm v ⇒ UnsoundDSIGNMAlgorithm v where

Unsound operations on DSIGNM sign keys. These operations violate secure forgetting constraints by leaking secrets to unprotected memory. Consider using the DirectSerialise / DirectDeserialise APIs instead.

Methods

rawSerialiseSignKeyDSIGNM ∷ (MonadST m, MonadThrow m) ⇒ SignKeyDSIGNM v → m ByteString

rawDeserialiseSignKeyDSIGNMWith ∷ (MonadST m, MonadThrow m) ⇒ MLockedAllocator m → ByteString → m (Maybe (SignKeyDSIGNM v))

encodeSignKeyDSIGNM ∷ (UnsoundDSIGNMAlgorithm v, MonadST m, MonadThrow m) ⇒ SignKeyDSIGNM v → m Encoding

decodeSignKeyDSIGNM ∷ ∀ m v s. (UnsoundDSIGNMAlgorithm v, MonadST m, MonadThrow m) ⇒ Decoder s (m (SignKeyDSIGNM v))

rawDeserialiseSignKeyDSIGNM ∷ (UnsoundDSIGNMAlgorithm v, MonadST m, MonadThrow m) ⇒ ByteString → m (Maybe (SignKeyDSIGNM v))