Cardano.Crypto.DSIGN.NeverUsed

Documentation

data NeverDSIGN

DSIGN never used

The type of keys and signatures is isomorphic to unit, but when actually trying to sign or verify something a runtime exception will be thrown.

data family VerKeyDSIGN v ∷ Type

data family SignKeyDSIGN v ∷ Type

data family SigDSIGN v ∷ Type