Cardano.Crypto.Hash.Class

Description

Abstract hashing functionality.

Documentation

class Typeable h ⇒ HashAlgorithm h where

Methods

hashAlgorithmName ∷ proxy h → String

sizeHash ∷ proxy h → Word

The size in bytes of the output of digest

digest ∷ proxy h → ByteString → ByteString

data ByteString

A space-efficient representation of a Word8 vector, supporting many efficient operations.

A ByteString contains 8-bit bytes, or by using the operations from Data.ByteString.Char8 it can be interpreted as containing 8-bit characters.

newtype Hash h a

Constructors

UnsafeHash ShortByteString

Core operations

hashWith ∷ ∀ h a. HashAlgorithm h ⇒ (a → ByteString) → a → Hash h a

Hash the given value, using a serialisation function to turn it into bytes.

hashWithSerialiser ∷ ∀ h a. HashAlgorithm h ⇒ (a → Encoding) → a → Hash h a

A variation on hashWith, but specially for CBOR encodings.

Conversions

castHash ∷ Hash h a → Hash h b

Cast the type of the hashed data.

The Hash type has a phantom type parameter to indicate what type the hash is of. It is sometimes necessary to fake this and hash a value of one type and use it where as hash of a different type is expected.

hashToBytes ∷ Hash h a → ByteString

The representation of the hash as bytes.

hashFromBytes ∷ ∀ h a. HashAlgorithm h ⇒ ByteString → Maybe (Hash h a)

Make a hash from it bytes representation.

It must be a a bytestring of the correct length, as given by sizeHash.

hashToBytesShort ∷ Hash h a → ShortByteString

The representation of the hash as bytes, as a ShortByteString.

hashFromBytesShort ∷ ∀ h a. HashAlgorithm h ⇒ ShortByteString → Maybe (Hash h a)

Make a hash from it bytes representation, as a ShortByteString.

It must be a a bytestring of the correct length, as given by sizeHash.

Rendering and parsing

hashToBytesAsHex ∷ Hash h a → ByteString

Convert the hash to hex encoding, as ByteString.

hashFromBytesAsHex ∷ HashAlgorithm h ⇒ ByteString → Maybe (Hash h a)

Make a hash from hex-encoded ByteString representation.

This can fail for the same reason as hashFromBytes, or because the input is invalid hex. The whole byte string must be valid hex, not just a prefix.

hashToTextAsHex ∷ Hash h a → Text

Convert the hash to hex encoding, as Text.

hashFromTextAsHex ∷ HashAlgorithm h ⇒ Text → Maybe (Hash h a)

Make a hash from hex-encoded Text representation.

This can fail for the same reason as hashFromBytes, or because the input is invalid hex. The whole byte string must be valid hex, not just a prefix.

hashToStringAsHex ∷ Hash h a → String

Convert the hash to hex encoding, as String.

hashFromStringAsHex ∷ HashAlgorithm h ⇒ String → Maybe (Hash h a)

Make a hash from hex-encoded String representation.

This can fail for the same reason as hashFromBytes, or because the input is invalid hex. The whole byte string must be valid hex, not just a prefix.

Other operations

xor ∷ Hash h a → Hash h a → Hash h a

XOR two hashes together TODO: fully deprecate this, or rename it and make it efficient.

Deprecated

hash ∷ ∀ h a. (HashAlgorithm h, ToCBOR a) ⇒ a → Hash h a

Deprecated: Use hashWith or hashWithSerialiser

fromHash ∷ Hash h a → Natural

Deprecated: Use bytesToNatural . hashToBytes

hashRaw ∷ ∀ h a. HashAlgorithm h ⇒ (a → ByteString) → a → Hash h a

Deprecated: Use hashWith

getHash ∷ Hash h a → ByteString

Deprecated: Use hashToBytes

getHashBytesAsHex ∷ Hash h a → ByteString

Deprecated: Use hashToBytesAsHex