Cardano.Crypto.KES.NeverUsed

Documentation

data NeverKES

KES never used

The type of keys and signatures is isomorphic to unit, but when actually trying to sign or verify something a runtime exception will be thrown.

data family VerKeyKES v ∷ Type

data family SignKeyKES v ∷ Type

data family SigKES v ∷ Type