Cardano.Crypto.KES.Single

Description

A standard signature scheme is a forward-secure signature scheme with a single time period.

This is the base case in the naive recursive implementation of the sum composition from section 3 of the "MMM" paper:

Composition and Efficiency Tradeoffs for Forward-Secure Digital Signatures By Tal Malkin, Daniele Micciancio and Sara Miner https://eprint.iacr.org/2001/034

Specfically it states:

In order to unify the presentation, we regard standard signature schemes
as forward-seure signature schemes with one time period, namely T = 1.

So this module simply provides a wrapper SingleKES that turns any DSIGNAlgorithm into an instance of KESAlgorithm with a single period.

See Cardano.Crypto.KES.Sum for the composition case.

Documentation

data SingleKES d

A standard signature scheme is a forward-secure signature scheme with a single time period.

data family VerKeyKES v :: Type

data family SignKeyKES v :: Type

data family SigKES v :: Type