Cardano.Crypto.KES.Sum

Description

A key evolving signatures implementation.

It is a naive recursive implementation of the sum composition from section 3.1 of the "MMM" paper:

Composition and Efficiency Tradeoffs for Forward-Secure Digital Signatures By Tal Malkin, Daniele Micciancio and Sara Miner https://eprint.iacr.org/2001/034

Specfically we do the binary sum composition directly as in the paper, and then use that in a nested/recursive fashion to construct a 7-level deep binary tree version.

This relies on Cardano.Crypto.KES.Single for the base case.

Documentation

data SumKES h d

A composition of two KES schemes to give a KES scheme with the sum of the time periods.

While we could do this with two independent KES schemes (i.e. two types) we only need it for two instances of the same scheme, and we save substantially on the size of the type and runtime dictionaries if we do it this way, especially when we start applying it recursively.

data family VerKeyKES v :: Type

data family SignKeyKES v :: Type

data family SigKES v :: Type

Type aliases for powers of binary sums

type Sum0KES d = SingleKES d

A 2^0 period KES

type Sum1KES d h = SumKES h (Sum0KES d)

A 2^1 period KES

type Sum2KES d h = SumKES h (Sum1KES d h)

A 2^2 period KES

type Sum3KES d h = SumKES h (Sum2KES d h)

A 2^3 period KES

type Sum4KES d h = SumKES h (Sum3KES d h)

A 2^4 period KES

type Sum5KES d h = SumKES h (Sum4KES d h)

A 2^5 period KES

type Sum6KES d h = SumKES h (Sum5KES d h)

A 2^6 period KES

type Sum7KES d h = SumKES h (Sum6KES d h)

A 2^7 period KES