Cardano.Crypto.Libsodium

Initialization

sodiumInit ∷ IO ()

MLocked memory management

data MLockedForeignPtr a

Foreign pointer to securely allocated memory.

withMLockedForeignPtr ∷ ∀ a b. MLockedForeignPtr a → (Ptr a → IO b) → IO b

allocMLockedForeignPtr ∷ Storable a ⇒ IO (MLockedForeignPtr a)

Allocate secure memory using c_sodium_malloc.

https://libsodium.gitbook.io/doc/memory_management

finalizeMLockedForeignPtr ∷ ∀ a. MLockedForeignPtr a → IO ()

traceMLockedForeignPtr ∷ (Storable a, Show a) ⇒ MLockedForeignPtr a → IO ()

Deprecated: Don't leave traceMLockedForeignPtr in production

MLocked bytes

data MLockedSizedBytes n

mlsbZero ∷ ∀ n. KnownNat n ⇒ MLockedSizedBytes n

Note: this doesn't need to allocate mlocked memory, but we do that for consistency

mlsbFromByteString ∷ ∀ n. KnownNat n ⇒ ByteString → MLockedSizedBytes n

mlsbFromByteStringCheck ∷ ∀ n. KnownNat n ⇒ ByteString → Maybe (MLockedSizedBytes n)

mlsbToByteString ∷ ∀ n. KnownNat n ⇒ MLockedSizedBytes n → ByteString

Note: the resulting ByteString will still refer to secure memory, but the types don't prevent it from be exposed.

mlsbFinalize ∷ MLockedSizedBytes n → IO ()

Calls finalizeMLockedForeignPtr on underlying pointer. This function invalidates argument.

Hashing

class HashAlgorithm h ⇒ SodiumHashAlgorithm h where

Methods

naclDigestPtr

Arguments

∷ proxy h

→ Ptr a
input

→ Int
input length

→ IO (MLockedSizedBytes (SizeHash h))

digestMLockedStorable ∷ ∀ h a proxy. (SodiumHashAlgorithm h, Storable a) ⇒ proxy h → Ptr a → MLockedSizedBytes (SizeHash h)

digestMLockedBS ∷ ∀ h proxy. SodiumHashAlgorithm h ⇒ proxy h → ByteString → MLockedSizedBytes (SizeHash h)

expandHash ∷ ∀ h proxy. SodiumHashAlgorithm h ⇒ proxy h → MLockedSizedBytes (SizeHash h) → (MLockedSizedBytes (SizeHash h), MLockedSizedBytes (SizeHash h))

Signing

class (DSIGNAlgorithm v, ContextDSIGN v ~ (), Signable v ~ SignableRepresentation) ⇒ SodiumDSIGNAlgorithm v where

Methods

naclSignDSIGNPtr ∷ Proxy v → Ptr a → Int → SodiumSignKeyDSIGN v → IO (SodiumSigDSIGN v)

naclVerifyDSIGNPtr ∷ Proxy v → SodiumVerKeyDSIGN v → Ptr a → Int → SodiumSigDSIGN v → IO (Either String ())

naclGenKeyDSIGN ∷ Proxy v → MLockedSizedBytes (SeedSizeDSIGN v) → SodiumSignKeyDSIGN v

naclDeriveVerKeyDSIGN ∷ Proxy v → SodiumSignKeyDSIGN v → SodiumVerKeyDSIGN v

naclSignDSIGN ∷ (SodiumDSIGNAlgorithm v, SignableRepresentation a) ⇒ Proxy v → a → SodiumSignKeyDSIGN v → SodiumSigDSIGN v

naclVerifyDSIGN ∷ (SodiumDSIGNAlgorithm v, SignableRepresentation a) ⇒ Proxy v → SodiumVerKeyDSIGN v → a → SodiumSigDSIGN v → Either String ()

naclForgetSignKeyDSIGN ∷ Proxy v → SodiumSignKeyDSIGN v → IO ()

type SodiumSignKeyDSIGN v = MLockedSizedBytes (SizeSignKeyDSIGN v)

type SodiumVerKeyDSIGN v = PinnedSizedBytes (SizeVerKeyDSIGN v)

type SodiumSigDSIGN v = PinnedSizedBytes (SizeSigDSIGN v)