Cardano.Crypto.Libsodium.C

Initialization

c_sodium_init :: IO Int

void sodium_init();

https://libsodium.gitbook.io/doc/usage

Memory management

c_sodium_memzero :: Ptr a -> CSize -> IO ()

void sodium_memzero(void * const pnt, const size_t len);

https://libsodium.gitbook.io/doc/memory_management#zeroing-memory

c_sodium_malloc :: CSize -> IO (Ptr a)

void *sodium_malloc(size_t size);

https://libsodium.gitbook.io/doc/memory_management

c_sodium_free :: Ptr a -> IO ()

void sodium_free(void *ptr);

https://libsodium.gitbook.io/doc/memory_management

c_sodium_free_funptr :: FunPtr (Ptr a -> IO ())

void sodium_free(void *ptr);

https://libsodium.gitbook.io/doc/memory_management

Hashing

SHA256

c_crypto_hash_sha256 :: SizedPtr CRYPTO_SHA256_BYTES -> Ptr CUChar -> CULLong -> IO Int

int crypto_hash_sha256(unsigned char *out, const unsigned char *in, unsigned long long inlen);

https://libsodium.gitbook.io/doc/advanced/sha-2_hash_function

c_crypto_hash_sha256_final :: SizedPtr CRYPTO_SHA256_STATE_SIZE -> SizedPtr CRYPTO_SHA256_BYTES -> IO Int

int crypto_hash_sha256_final(crypto_hash_sha256_state *state, unsigned char *out);

c_crypto_hash_sha256_init :: SizedPtr CRYPTO_SHA256_STATE_SIZE -> IO Int

int crypto_hash_sha256_init(crypto_hash_sha256_state *state);

c_crypto_hash_sha256_update :: SizedPtr CRYPTO_SHA256_STATE_SIZE -> Ptr CUChar -> CULLong -> IO Int

int crypto_hash_sha256_update(crypto_hash_sha256_state *state, const unsigned char *in, unsigned long long inlen);

Blake2b 256

c_crypto_generichash_blake2b :: Ptr out -> CSize -> Ptr CUChar -> CULLong -> Ptr key -> CSize -> IO Int

int crypto_generichash_blake2b(unsigned char *out, size_t outlen, const unsigned char *in, unsigned long long inlen, const unsigned char *key, size_t keylen);

https://libsodium.gitbook.io/doc/hashing/generic_hashing

c_crypto_generichash_blake2b_final :: SizedPtr CRYPTO_BLAKE2B_256_STATE_SIZE -> Ptr out -> CSize -> IO Int

int crypto_generichash_blake2b_final(crypto_generichash_blake2b_state *state, unsigned char *out, const size_t outlen);

c_crypto_generichash_blake2b_init :: SizedPtr CRYPTO_BLAKE2B_256_STATE_SIZE -> Ptr key -> CSize -> CSize -> IO Int

int crypto_generichash_blake2b_init(crypto_generichash_blake2b_state *state, const unsigned char *key, const size_t keylen, const size_t outlen);

c_crypto_generichash_blake2b_update :: SizedPtr CRYPTO_BLAKE2B_256_STATE_SIZE -> Ptr CUChar -> CULLong -> IO Int

int crypto_generichash_blake2b_update(crypto_generichash_blake2b_state *state, const unsigned char *in, unsigned long long inlen);

ED25519

c_crypto_sign_ed25519_seed_keypair :: SizedPtr CRYPTO_SIGN_ED25519_PUBLICKEYBYTES -> SizedPtr CRYPTO_SIGN_ED25519_SECRETKEYBYTES -> SizedPtr CRYPTO_SIGN_ED25519_SEEDBYTES -> IO Int

int crypto_sign_ed25519_seed_keypair(unsigned char *pk, unsigned char *sk, const unsigned char *seed);

c_crypto_sign_ed25519_sk_to_seed :: SizedPtr CRYPTO_SIGN_ED25519_SEEDBYTES -> SizedPtr CRYPTO_SIGN_ED25519_SECRETKEYBYTES -> IO Int

int crypto_sign_ed25519_sk_to_seed(unsigned char *seed, const unsigned char *sk);

c_crypto_sign_ed25519_detached :: SizedPtr CRYPTO_SIGN_ED25519_BYTES -> Ptr CULLong -> Ptr CUChar -> CULLong -> SizedPtr CRYPTO_SIGN_ED25519_SECRETKEYBYTES -> IO Int

int crypto_sign_ed25519_detached(unsigned char *sig, unsigned long long *siglen_p, const unsigned char *m, unsigned long long mlen, const unsigned char *sk);

c_crypto_sign_ed25519_verify_detached :: SizedPtr CRYPTO_SIGN_ED25519_BYTES -> Ptr CUChar -> CULLong -> SizedPtr CRYPTO_SIGN_ED25519_PUBLICKEYBYTES -> IO Int

int crypto_sign_ed25519_verify_detached(const unsigned char *sig, const unsigned char *m, unsigned long long mlen, const unsigned char *pk);

c_crypto_sign_ed25519_sk_to_pk :: SizedPtr CRYPTO_SIGN_ED25519_PUBLICKEYBYTES -> SizedPtr CRYPTO_SIGN_ED25519_SECRETKEYBYTES -> IO Int

int crypto_sign_ed25519_sk_to_pk(unsigned char *pk, const unsigned char *sk);

Helpers

c_sodium_compare :: Ptr a -> Ptr a -> CSize -> IO Int

int sodium_compare(const void * const b1_, const void * const b2_, size_t len);

https://libsodium.gitbook.io/doc/helpers#comparing-large-numbers

Constants

type CRYPTO_SHA256_BYTES = 32

type CRYPTO_SHA512_BYTES = 64

type CRYPTO_BLAKE2B_256_BYTES = 32

type CRYPTO_SHA256_STATE_SIZE = 104

type CRYPTO_SHA512_STATE_SIZE = 208

type CRYPTO_BLAKE2B_256_STATE_SIZE = 384

type CRYPTO_SIGN_ED25519_BYTES = 64

type CRYPTO_SIGN_ED25519_SEEDBYTES = 32

type CRYPTO_SIGN_ED25519_PUBLICKEYBYTES = 32

type CRYPTO_SIGN_ED25519_SECRETKEYBYTES = 64