Cardano.Crypto.Libsodium.MLockedBytes

Documentation

data MLockedSizedBytes (n :: Nat)

A block of raw memory of a known size, protected with mlock().

data SizedVoid (n :: Nat)

A void type with a type-level size attached to it. We need this in order to express "pointer to a block of memory of a particular size that can be manipulated through the pointer, but not as a plain Haskell value" as Ptr (SizedVoid n), or ForeignPtr (SizedVoid n), or MLockedForeignPtr (SizedVoid n).

withMLSB :: forall b n m. MonadSodium m => MLockedSizedBytes n -> (Ptr (SizedVoid n) -> m b) -> m b

withMLSBChunk :: forall b n n' m. (MonadSodium m, MonadST m, KnownNat n, KnownNat n') => MLockedSizedBytes n -> Int -> (MLockedSizedBytes n' -> m b) -> m b

mlsbNew :: forall n m. (KnownNat n, MonadSodium m) => m (MLockedSizedBytes n)

Allocate a new MLockedSizedBytes. The caller is responsible for deallocating it (mlsbFinalize) when done with it. The contents of the memory block is undefined.

mlsbNewZero :: forall n m. (KnownNat n, MonadSodium m) => m (MLockedSizedBytes n)

Allocate a new MLockedSizedBytes, and pre-fill it with zeroes. The caller is responsible for deallocating it (mlsbFinalize) when done with it. (See also mlsbNew).

mlsbZero :: forall n m. (KnownNat n, MonadSodium m) => MLockedSizedBytes n -> m ()

Overwrite an existing MLockedSizedBytes with zeroes.

mlsbFromByteString :: forall n m. (KnownNat n, MonadSodium m, MonadST m) => ByteString -> m (MLockedSizedBytes n)

Allocate a new MLockedSizedBytes, and fill it with the contents of a ByteString. The size of the input is not checked. Note: since the input ByteString is a plain old Haskell value, it has already violated the secure-forgetting properties afforded by MLockedSizedBytes, so this function is useless outside of testing. Use mlsbNew or mlsbNewZero to create MLockedSizedBytes values, and manipulate them through withMLSB, mlsbUseAsCPtr, or mlsbUseAsSizedPtr. (See also mlsbFromByteStringCheck)

mlsbFromByteStringCheck :: forall n m. (KnownNat n, MonadSodium m, MonadST m) => ByteString -> m (Maybe (MLockedSizedBytes n))

Allocate a new MLockedSizedBytes, and fill it with the contents of a ByteString. The size of the input is checked. Note: since the input ByteString is a plain old Haskell value, it has already violated the secure-forgetting properties afforded by MLockedSizedBytes, so this function is useless outside of testing. Use mlsbNew or mlsbNewZero to create MLockedSizedBytes values, and manipulate them through withMLSB, mlsbUseAsCPtr, or mlsbUseAsSizedPtr. (See also mlsbFromByteString)

mlsbAsByteString :: forall n. KnownNat n => MLockedSizedBytes n -> ByteString

Note: the resulting ByteString will still refer to secure memory, but the types don't prevent it from be exposed. Note further that any subsequent operations (splicing & dicing, copying, conversion, packing/unpacking, etc.) on the resulting ByteString may create copies of the mlocked memory on the unprotected GHC heap, and thus leak secrets, so use this function with extreme care.

mlsbToByteString :: forall n m. (KnownNat n, MonadSodium m, MonadST m) => MLockedSizedBytes n -> m ByteString

Note: this function will leak mlocked memory to the Haskell heap and should not be used in production code.

mlsbUseAsCPtr :: MonadSodium m => MLockedSizedBytes n -> (Ptr Word8 -> m r) -> m r

Use an MLockedSizedBytes value as a raw C pointer. Care should be taken to never copy the contents of the MLockedSizedBytes value into managed memory through the raw pointer, because that would violate the secure-forgetting property of mlocked memory.

mlsbUseAsSizedPtr :: forall n r m. MonadSodium m => MLockedSizedBytes n -> (SizedPtr n -> m r) -> m r

Use an MLockedSizedBytes value as a SizedPtr of the same size. Care should be taken to never copy the contents of the MLockedSizedBytes value into managed memory through the sized pointer, because that would violate the secure-forgetting property of mlocked memory.

mlsbFinalize :: MonadSodium m => MLockedSizedBytes n -> m ()

Calls finalizeMLockedForeignPtr on underlying pointer. This function invalidates argument.

mlsbCopy :: forall n m. (KnownNat n, MonadSodium m) => MLockedSizedBytes n -> m (MLockedSizedBytes n)

Create a deep mlocked copy of an MLockedSizedBytes.

traceMLSB :: KnownNat n => MLockedSizedBytes n -> IO ()

Deprecated: Don't leave traceMLockedForeignPtr in production

mlsbCompare :: forall n m. (MonadSodium m, MonadST m, KnownNat n) => MLockedSizedBytes n -> MLockedSizedBytes n -> m Ordering

compareM on MLockedSizedBytes

mlsbEq :: forall n m. (MonadSodium m, MonadST m, KnownNat n) => MLockedSizedBytes n -> MLockedSizedBytes n -> m Bool

equalsM on MLockedSizedBytes