Cardano.Crypto.Libsodium.MLockedSeed

Documentation

newtype MLockedSeed n

A seed of size n, stored in mlocked memory. This is required to prevent the seed from leaking to disk via swapping and reclaiming or scanning memory after its content has been moved.

Constructors

MLockedSeed

Fields

mlockedSeedMLSB ∷ MLockedSizedBytes n

withMLockedSeedAsMLSB ∷ Functor m ⇒ (MLockedSizedBytes n → m (MLockedSizedBytes n)) → MLockedSeed n → m (MLockedSeed n)

mlockedSeedCopy ∷ (KnownNat n, MonadST m) ⇒ MLockedSeed n → m (MLockedSeed n)

mlockedSeedCopyWith ∷ (KnownNat n, MonadST m) ⇒ MLockedAllocator m → MLockedSeed n → m (MLockedSeed n)

mlockedSeedNew ∷ (KnownNat n, MonadST m) ⇒ m (MLockedSeed n)

mlockedSeedNewWith ∷ (KnownNat n, MonadST m) ⇒ MLockedAllocator m → m (MLockedSeed n)

mlockedSeedNewZero ∷ (KnownNat n, MonadST m) ⇒ m (MLockedSeed n)

mlockedSeedNewZeroWith ∷ (KnownNat n, MonadST m) ⇒ MLockedAllocator m → m (MLockedSeed n)

mlockedSeedNewRandom ∷ ∀ n. KnownNat n ⇒ IO (MLockedSeed n)

mlockedSeedNewRandomWith ∷ ∀ n. KnownNat n ⇒ MLockedAllocator IO → IO (MLockedSeed n)

mlockedSeedFinalize ∷ MonadST m ⇒ MLockedSeed n → m ()

mlockedSeedUseAsCPtr ∷ MonadST m ⇒ MLockedSeed n → (Ptr Word8 → m b) → m b

mlockedSeedUseAsSizedPtr ∷ MonadST m ⇒ MLockedSeed n → (SizedPtr n → m b) → m b