Cardano.Crypto.Libsodium.Memory

High-level memory management

data MLockedForeignPtr a

Foreign pointer to securely allocated memory.

withMLockedForeignPtr ∷ MonadST m ⇒ MLockedForeignPtr a → (Ptr a → m b) → m b

finalizeMLockedForeignPtr ∷ MonadST m ⇒ MLockedForeignPtr a → m ()

traceMLockedForeignPtr ∷ (Storable a, Show a, MonadST m) ⇒ MLockedForeignPtr a → m ()

Warning: Do not use traceMLockedForeignPtr in production

MLocked allocations

mlockedMalloc ∷ MonadST m ⇒ MLockedAllocator m

newtype MLockedAllocator m

Constructors

MLockedAllocator

Fields

mlAllocate ∷ ∀ a. CSize → m (MLockedForeignPtr a)

mlockedAlloca ∷ ∀ a b m. (MonadST m, MonadThrow m) ⇒ CSize → (Ptr a → m b) → m b

mlockedAllocaSized ∷ ∀ m n b. (MonadST m, MonadThrow m, KnownNat n) ⇒ (SizedPtr n → m b) → m b

mlockedAllocForeignPtr ∷ ∀ a m. (MonadST m, Storable a) ⇒ m (MLockedForeignPtr a)

mlockedAllocForeignPtrBytes ∷ MonadST m ⇒ CSize → CSize → m (MLockedForeignPtr a)

Allocations using an explicit allocator

mlockedAllocaWith ∷ ∀ a b m. (MonadThrow m, MonadST m) ⇒ MLockedAllocator m → CSize → (Ptr a → m b) → m b

mlockedAllocaSizedWith ∷ ∀ m n b. (MonadST m, MonadThrow m, KnownNat n) ⇒ MLockedAllocator m → (SizedPtr n → m b) → m b

mlockedAllocForeignPtrWith ∷ ∀ a m. Storable a ⇒ MLockedAllocator m → m (MLockedForeignPtr a)

mlockedAllocForeignPtrBytesWith ∷ MLockedAllocator m → CSize → CSize → m (MLockedForeignPtr a)

Unmanaged memory, generalized to MonadST

zeroMem ∷ MonadST m ⇒ Ptr a → CSize → m ()

copyMem ∷ MonadST m ⇒ Ptr a → Ptr a → CSize → m ()

allocaBytes ∷ (MonadThrow m, MonadST m) ⇒ Int → (Ptr a → m b) → m b

ForeignPtr operations, generalized to MonadST

newtype ForeignPtr (m ∷ Type → Type) a

A ForeignPtr type, generalized to MonadST. The type is tagged with the correct Monad m in order to ensure that foreign pointers created in one ST context can only be used within the same ST context.

Constructors

ForeignPtr

Fields

unsafeRawForeignPtr ∷ ForeignPtr a

mallocForeignPtrBytes ∷ MonadST m ⇒ Int → m (ForeignPtr m a)

withForeignPtr ∷ MonadST m ⇒ ForeignPtr m a → (Ptr a → m b) → m b

withForeignPtr, generalized to MonadST. Caveat: if the monadic action passed to withForeignPtr does not terminate (e.g., forever), the ForeignPtr finalizer may run prematurely.

ByteString memory access, generalized to MonadST

unpackByteStringCStringLen ∷ (MonadThrow m, MonadST m) ⇒ ByteString → (CStringLen → m a) → m a

Unpacks a ByteString into a temporary buffer and runs the provided ST function on it.

packByteStringCStringLen ∷ MonadST m ⇒ CStringLen → m ByteString