Cardano.Crypto.MEqOrd

Documentation

class MEq m a where

Monadic flavor of Eq, for things that can only be compared in a monadic context. This is needed because we cannot have a sound Eq instance on mlocked memory types.

Methods

equalsM :: a -> a -> m Bool

nequalsM :: (Functor m, MEq m a) => a -> a -> m Bool

(==!) :: MEq m a => a -> a -> m Bool infix 4

Infix version of equalsM

(!=!) :: (Functor m, MEq m a) => a -> a -> m Bool infix 4

Infix version of nequalsM

newtype PureMEq a

Helper newtype, useful for defining MEq in terms of Eq for types that have sound Eq instances, using DerivingVia. An Applicative context must be provided for such instances to work, so this will generally require StandaloneDeriving as well.

Ex.: deriving via PureEq Int instance Applicative m => MEq m Int

Constructors

PureMEq a