Cardano.Crypto.Util

Documentation

class Empty a

class SignableRepresentation a where

A class of types that have a representation in bytes that can be used for signing and verifying.

Methods

getSignableRepresentation :: a -> ByteString

getRandomWord64 :: MonadRandom m => m Word64

Simple serialisation used in mock instances

readBinaryWord64 :: ByteString -> Word64

writeBinaryWord64 :: Word64 -> ByteString

readBinaryNatural :: ByteString -> Natural

writeBinaryNatural :: Int -> Natural -> ByteString

splitsAt :: [Int] -> ByteString -> [ByteString]

Low level conversions

bytesToNatural :: ByteString -> Natural

Create a Natural out of a ByteString, in big endian.

This is fast enough to use in production.

naturalToBytes :: Int -> Natural -> ByteString

The inverse of bytesToNatural. Note that this is a naive implementation and only suitable for tests.

ByteString manipulation

slice :: Word -> Word -> ByteString -> ByteString

Base16 conversion

decodeHexByteString :: ByteString -> Int -> Either String ByteString

Decode base16 ByteString, while ensuring expected length.

decodeHexString :: String -> Int -> Either String ByteString

Decode base16 String, while ensuring expected length. Unlike decodeHexByteString this function expects a '0x' prefix.

decodeHexStringQ :: String -> Int -> Q Exp

Decode a String with Hex characters, while ensuring expected length.