Cardano.Crypto.VRF.Class

Description

Abstract Verifiable Random Functions.

VRF algorithm class

class (Typeable v, Show (VerKeyVRF v), Eq (VerKeyVRF v), Show (SignKeyVRF v), Show (CertVRF v), Eq (CertVRF v), NoThunks (CertVRF v), NoThunks (VerKeyVRF v), NoThunks (SignKeyVRF v)) ⇒ VRFAlgorithm v where

Minimal complete definition

algorithmNameVRF, deriveVerKeyVRF, evalVRF, verifyVRF, seedSizeVRF, (genKeyVRF | genKeyPairVRF), rawSerialiseVerKeyVRF, rawSerialiseSignKeyVRF, rawSerialiseCertVRF, rawDeserialiseVerKeyVRF, rawDeserialiseSignKeyVRF, rawDeserialiseCertVRF, sizeVerKeyVRF, sizeSignKeyVRF, sizeCertVRF, sizeOutputVRF

Associated Types

data VerKeyVRF v ∷ Type

data SignKeyVRF v ∷ Type

data CertVRF v ∷ Type

type ContextVRF v ∷ Type

Context required to run the VRF algorithm

Unit by default (no context required)

type ContextVRF v = ()

type Signable v ∷ Type → Constraint

type Signable c = Empty

Methods

algorithmNameVRF ∷ proxy v → String

deriveVerKeyVRF ∷ SignKeyVRF v → VerKeyVRF v

hashVerKeyVRF ∷ HashAlgorithm h ⇒ VerKeyVRF v → Hash h (VerKeyVRF v)

evalVRF ∷ (HasCallStack, Signable v a) ⇒ ContextVRF v → a → SignKeyVRF v → (OutputVRF v, CertVRF v)

verifyVRF ∷ (HasCallStack, Signable v a) ⇒ ContextVRF v → VerKeyVRF v → a → CertVRF v → Maybe (OutputVRF v)

genKeyVRF ∷ Seed → SignKeyVRF v

genKeyPairVRF ∷ Seed → (SignKeyVRF v, VerKeyVRF v)

seedSizeVRF ∷ proxy v → Word

The upper bound on the Seed size needed by genKeyVRF, in bytes.

sizeVerKeyVRF ∷ proxy v → Word

sizeSignKeyVRF ∷ proxy v → Word

sizeCertVRF ∷ proxy v → Word

sizeOutputVRF ∷ proxy v → Word

rawSerialiseVerKeyVRF ∷ VerKeyVRF v → ByteString

rawSerialiseSignKeyVRF ∷ SignKeyVRF v → ByteString

rawSerialiseCertVRF ∷ CertVRF v → ByteString

rawDeserialiseVerKeyVRF ∷ ByteString → Maybe (VerKeyVRF v)

rawDeserialiseSignKeyVRF ∷ ByteString → Maybe (SignKeyVRF v)

rawDeserialiseCertVRF ∷ ByteString → Maybe (CertVRF v)

VRF output

newtype OutputVRF v

The output bytes of the VRF.

The output size is a fixed number of bytes and is given by sizeOutputVRF.

Constructors

OutputVRF

Fields

getOutputVRFBytes ∷ ByteString

getOutputVRFNatural ∷ OutputVRF v → Natural

The output bytes of the VRF interpreted as a big endian natural number.

The range of this number is determined by the size of the VRF output bytes. It is thus in the range 0 .. 2 ^ (8 * sizeOutputVRF proxy) - 1.

mkTestOutputVRF ∷ ∀ v. VRFAlgorithm v ⇒ Natural → OutputVRF v

For testing purposes, make an OutputVRF from a Natural.

The OutputVRF will be of the appropriate size for the VRFAlgorithm.

CertifiedVRF wrapper

data CertifiedVRF v a

Constructors

CertifiedVRF

Fields

certifiedOutput ∷ !(OutputVRF v)

certifiedProof ∷ !(CertVRF v)

evalCertified ∷ (VRFAlgorithm v, Signable v a) ⇒ ContextVRF v → a → SignKeyVRF v → CertifiedVRF v a

verifyCertified ∷ (VRFAlgorithm v, Signable v a) ⇒ ContextVRF v → VerKeyVRF v → a → CertifiedVRF v a → Bool

CBOR encoding and decoding

encodeVerKeyVRF ∷ VRFAlgorithm v ⇒ VerKeyVRF v → Encoding

decodeVerKeyVRF ∷ ∀ v s. VRFAlgorithm v ⇒ Decoder s (VerKeyVRF v)

encodeSignKeyVRF ∷ VRFAlgorithm v ⇒ SignKeyVRF v → Encoding

decodeSignKeyVRF ∷ ∀ v s. VRFAlgorithm v ⇒ Decoder s (SignKeyVRF v)

encodeCertVRF ∷ VRFAlgorithm v ⇒ CertVRF v → Encoding

decodeCertVRF ∷ ∀ v s. VRFAlgorithm v ⇒ Decoder s (CertVRF v)

Encoded Size expressions

encodedVerKeyVRFSizeExpr ∷ ∀ v. VRFAlgorithm v ⇒ Proxy (VerKeyVRF v) → Size

Size expression for VerKeyVRF which is using sizeVerKeyVRF encoded as Size.

encodedSignKeyVRFSizeExpr ∷ ∀ v. VRFAlgorithm v ⇒ Proxy (SignKeyVRF v) → Size

Size expression for SignKeyVRF which is using sizeSignKeyVRF encoded as Size

encodedCertVRFSizeExpr ∷ ∀ v. VRFAlgorithm v ⇒ Proxy (CertVRF v) → Size

Size expression for CertVRF which is using sizeCertVRF encoded as Size.