Cardano.Crypto.VRF.NeverUsed

Documentation

data NeverVRF

VRF not available

The type of keys and certificates is isomorphic to unit, but when actually trying to sign or verify something a runtime exception will be thrown.

data family VerKeyVRF v ∷ Type

data family SignKeyVRF v ∷ Type

data family CertVRF v ∷ Type