Cardano.Crypto.DSIGN.Class

Description

Abstract digital signatures.

Documentation

class (Typeable v, Show (VerKeyDSIGN v), Eq (VerKeyDSIGN v), Show (SignKeyDSIGN v), Show (SigDSIGN v), Eq (SigDSIGN v), NoUnexpectedThunks (SigDSIGN v), NoUnexpectedThunks (SignKeyDSIGN v), NoUnexpectedThunks (VerKeyDSIGN v)) ⇒ DSIGNAlgorithm v where

Minimal complete definition

algorithmNameDSIGN, deriveVerKeyDSIGN, abstractSizeVKey, abstractSizeSig, signDSIGN, verifyDSIGN, genKeyDSIGN, seedSizeDSIGN, (rawSerialiseVerKeyDSIGN | encodeVerKeyDSIGN), (rawSerialiseSignKeyDSIGN | encodeSignKeyDSIGN), (rawSerialiseSigDSIGN | encodeSigDSIGN), (rawDeserialiseVerKeyDSIGN | decodeVerKeyDSIGN), (rawDeserialiseSignKeyDSIGN | decodeSignKeyDSIGN), (rawDeserialiseSigDSIGN | decodeSigDSIGN)

Associated Types

data VerKeyDSIGN v ∷ Type

data SignKeyDSIGN v ∷ Type

data SigDSIGN v ∷ Type

type ContextDSIGN v ∷ Type

Context required to run the DSIGN algorithm

Unit by default (no context required)

type Signable v ∷ Type → Constraint

Methods

algorithmNameDSIGN ∷ proxy v → String

deriveVerKeyDSIGN ∷ SignKeyDSIGN v → VerKeyDSIGN v

hashVerKeyDSIGN ∷ HashAlgorithm h ⇒ VerKeyDSIGN v → Hash h (VerKeyDSIGN v)

abstractSizeVKey ∷ proxy v → Natural

Abstract sizes for verification keys and signatures, specifies an upper bound on the real byte sizes.

abstractSizeSig ∷ proxy v → Natural

signDSIGN ∷ (Signable v a, HasCallStack) ⇒ ContextDSIGN v → a → SignKeyDSIGN v → SigDSIGN v

verifyDSIGN ∷ (Signable v a, HasCallStack) ⇒ ContextDSIGN v → VerKeyDSIGN v → a → SigDSIGN v → Either String ()

genKeyDSIGN ∷ Seed → SignKeyDSIGN v

seedSizeDSIGN ∷ proxy v → Natural

The upper bound on the Seed size needed by genKeyDSIGN

rawSerialiseVerKeyDSIGN ∷ VerKeyDSIGN v → ByteString

rawSerialiseSignKeyDSIGN ∷ SignKeyDSIGN v → ByteString

rawSerialiseSigDSIGN ∷ SigDSIGN v → ByteString

rawDeserialiseVerKeyDSIGN ∷ ByteString → Maybe (VerKeyDSIGN v)

rawDeserialiseSignKeyDSIGN ∷ ByteString → Maybe (SignKeyDSIGN v)

rawDeserialiseSigDSIGN ∷ ByteString → Maybe (SigDSIGN v)

encodeVerKeyDSIGN ∷ VerKeyDSIGN v → Encoding

encodeSignKeyDSIGN ∷ SignKeyDSIGN v → Encoding

encodeSigDSIGN ∷ SigDSIGN v → Encoding

decodeVerKeyDSIGN ∷ Decoder s (VerKeyDSIGN v)

decodeSignKeyDSIGN ∷ Decoder s (SignKeyDSIGN v)

decodeSigDSIGN ∷ Decoder s (SigDSIGN v)

newtype SignedDSIGN v a

Constructors

SignedDSIGN (SigDSIGN v)

signedDSIGN ∷ (DSIGNAlgorithm v, Signable v a) ⇒ ContextDSIGN v → a → SignKeyDSIGN v → SignedDSIGN v a

verifySignedDSIGN ∷ (DSIGNAlgorithm v, Signable v a, HasCallStack) ⇒ ContextDSIGN v → VerKeyDSIGN v → a → SignedDSIGN v a → Either String ()

encodeSignedDSIGN ∷ DSIGNAlgorithm v ⇒ SignedDSIGN v a → Encoding

decodeSignedDSIGN ∷ DSIGNAlgorithm v ⇒ Decoder s (SignedDSIGN v a)

data Seed

A seed contains a finite number of bytes, and is used for seeding cryptographic algorithms including key generation.

This is not itself a PRNG, but can be used to seed a PRNG.