Cardano.Crypto.Hash.Class

Description

Abstract hashing functionality.

Documentation

class (KnownNat (SizeHash h), Typeable h) => HashAlgorithm h where

Associated Types

type SizeHash h :: Nat

Size of hash digest

Methods

hashAlgorithmName :: proxy h -> String

digest :: proxy h -> ByteString -> ByteString

sizeHash :: forall h proxy. HashAlgorithm h => proxy h -> Word

The size in bytes of the output of digest

data ByteString

A space-efficient representation of a Word8 vector, supporting many efficient operations.

A ByteString contains 8-bit bytes, or by using the operations from Data.ByteString.Char8 it can be interpreted as containing 8-bit characters.

data Hash h a where

Bundled Patterns

pattern UnsafeHash :: forall h a. HashAlgorithm h => ShortByteString -> Hash h a

data PackedBytes (n :: Nat) where

Constructors

PackedBytes8 :: !Word64 -> PackedBytes 8

PackedBytes28 :: !Word64 -> !Word64 -> !Word64 -> !Word32 -> PackedBytes 28

PackedBytes32 :: !Word64 -> !Word64 -> !Word64 -> !Word64 -> PackedBytes 32

Core operations

hashWith :: forall h a. HashAlgorithm h => (a -> ByteString) -> a -> Hash h a

Hash the given value, using a serialisation function to turn it into bytes.

hashWithSerialiser :: forall h a. HashAlgorithm h => (a -> Encoding) -> a -> Hash h a

A variation on hashWith, but specially for CBOR encodings.

Conversions

castHash :: Hash h a -> Hash h b

Cast the type of the hashed data.

The Hash type has a phantom type parameter to indicate what type the hash is of. It is sometimes necessary to fake this and hash a value of one type and use it where as hash of a different type is expected.

hashToBytes :: Hash h a -> ByteString

The representation of the hash as bytes.

hashFromBytes

Arguments

:: forall h a. HashAlgorithm h

=> ByteString
It must have an exact length, as given by sizeHash.

-> Maybe (Hash h a)

Make a hash from it bytes representation.

hashToBytesShort :: Hash h a -> ShortByteString

The representation of the hash as bytes, as a ShortByteString.

hashFromBytesShort

Arguments

:: forall h a. HashAlgorithm h

=> ShortByteString
It must be a buffer of exact length, as given by sizeHash.

-> Maybe (Hash h a)

Make a hash from it bytes representation, as a ShortByteString.

hashFromOffsetBytesShort

Arguments

:: forall h a. HashAlgorithm h

=> ShortByteString
It must be a buffer that contains at least sizeHash many bytes staring at an offset.

-> Int
Offset in number of bytes

-> Maybe (Hash h a)

Just like hashFromBytesShort, but allows using a region of a ShortByteString.

hashToPackedBytes :: Hash h a -> PackedBytes (SizeHash h)

O(1) - Get the underlying hash representation

hashFromPackedBytes :: PackedBytes (SizeHash h) -> Hash h a

O(1) - Construct hash from the underlying representation

Rendering and parsing

hashToBytesAsHex :: Hash h a -> ByteString

Convert the hash to hex encoding, as ByteString.

hashFromBytesAsHex :: HashAlgorithm h => ByteString -> Maybe (Hash h a)

Make a hash from hex-encoded ByteString representation.

This can fail for the same reason as hashFromBytes, or because the input is invalid hex. The whole byte string must be valid hex, not just a prefix.

hashToTextAsHex :: Hash h a -> Text

Convert the hash to hex encoding, as Text.

hashFromTextAsHex :: HashAlgorithm h => Text -> Maybe (Hash h a)

Make a hash from hex-encoded Text representation.

This can fail for the same reason as hashFromBytes, or because the input is invalid hex. The whole byte string must be valid hex, not just a prefix.

hashToStringAsHex :: Hash h a -> String

Convert the hash to hex encoding, as String.

hashFromStringAsHex :: HashAlgorithm h => String -> Maybe (Hash h a)

Make a hash from hex-encoded String representation.

This can fail for the same reason as hashFromBytes, or because the input is invalid hex. The whole byte string must be valid hex, not just a prefix.

Other operations

xor :: Hash h a -> Hash h a -> Hash h a

XOR two hashes together

Deprecated

hash :: forall h a. (HashAlgorithm h, ToCBOR a) => a -> Hash h a

Deprecated: Use hashWith or hashWithSerialiser

fromHash :: Hash h a -> Natural

Deprecated: Use bytesToNatural . hashToBytes

hashRaw :: forall h a. HashAlgorithm h => (a -> ByteString) -> a -> Hash h a

Deprecated: Use hashWith

getHash :: Hash h a -> ByteString

Deprecated: Use hashToBytes

getHashBytesAsHex :: Hash h a -> ByteString

Deprecated: Use hashToBytesAsHex