Cardano.Crypto.Init

Description

Initialization for the library's functionality

Documentation

cryptoInit :: IO ()

Initialize all the functionality provided by this library. This should be called at least once before you use anything this library provides, in main.

It is safe to call this multiple times, but isn't necessary.

Note

This includes a call to sodiumInit.