Cardano.Crypto.KES.Class

Description

Abstract key evolving signatures.

KES algorithm class

class (Typeable v, Show (VerKeyKES v), Eq (VerKeyKES v), Show (SigKES v), Eq (SigKES v), NoThunks (SigKES v), NoThunks (SignKeyKES v), NoThunks (VerKeyKES v), KnownNat (SeedSizeKES v), KnownNat (SizeVerKeyKES v), KnownNat (SizeSignKeyKES v), KnownNat (SizeSigKES v)) ⇒ KESAlgorithm v where

Minimal complete definition

algorithmNameKES, verifyKES, totalPeriodsKES, rawSerialiseVerKeyKES, rawSerialiseSigKES, rawDeserialiseVerKeyKES, rawDeserialiseSigKES, deriveVerKeyKES, signKES, updateKESWith, genKeyKESWith, forgetSignKeyKESWith

Associated Types

data VerKeyKES v ∷ Type

data SigKES v ∷ Type

data SignKeyKES v ∷ Type

type SeedSizeKES v ∷ Nat

type SizeVerKeyKES v ∷ Nat

type SizeSignKeyKES v ∷ Nat

type SizeSigKES v ∷ Nat

type ContextKES v ∷ Type

Context required to run the KES algorithm

Unit by default (no context required)

type ContextKES v = ()

type Signable v ∷ Type → Constraint

type Signable v = Empty

Methods

algorithmNameKES ∷ proxy v → String

hashVerKeyKES ∷ HashAlgorithm h ⇒ VerKeyKES v → Hash h (VerKeyKES v)

verifyKES

Arguments

∷ (Signable v a, HasCallStack)

⇒ ContextKES v

→ VerKeyKES v

→ Period
The current period for the key

→ a

→ SigKES v

→ Either String ()

Full KES verification. This method checks that the signature itself checks out (as per verifySigKES), and also makes sure that it matches the provided VerKey.

totalPeriodsKES ∷ proxy v → Word

Return the total number of KES periods supported by this algorithm. The KES algorithm is assumed to support a fixed maximum number of periods, not a variable number.

Do note that this is the total number of periods not the total number of evolutions. The difference is off-by-one. For example if there are 2 periods (period 0 and 1) then there is only one evolution.

rawSerialiseVerKeyKES ∷ VerKeyKES v → ByteString

rawSerialiseSigKES ∷ SigKES v → ByteString

rawDeserialiseVerKeyKES ∷ ByteString → Maybe (VerKeyKES v)

rawDeserialiseSigKES ∷ ByteString → Maybe (SigKES v)

deriveVerKeyKES ∷ (MonadST m, MonadThrow m) ⇒ SignKeyKES v → m (VerKeyKES v)

signKES

Arguments

∷ ∀ a m. (Signable v a, MonadST m, MonadThrow m)

⇒ ContextKES v

→ Period
The current period for the key

→ a

→ SignKeyKES v

→ m (SigKES v)

updateKESWith

Arguments

∷ (MonadST m, MonadThrow m)

⇒ MLockedAllocator m

→ ContextKES v

→ SignKeyKES v

→ Period
The current period for the key, not the target period.

→ m (Maybe (SignKeyKES v))

genKeyKESWith ∷ (MonadST m, MonadThrow m) ⇒ MLockedAllocator m → MLockedSeed (SeedSizeKES v) → m (SignKeyKES v)

forgetSignKeyKESWith ∷ (MonadST m, MonadThrow m) ⇒ MLockedAllocator m → SignKeyKES v → m ()

Forget a signing key synchronously, rather than waiting for GC. In some non-mock instances this provides a guarantee that the signing key is no longer in memory.

The precondition is that this key value will not be used again.

genKeyKES ∷ ∀ v m. (KESAlgorithm v, MonadST m, MonadThrow m) ⇒ MLockedSeed (SeedSizeKES v) → m (SignKeyKES v)

Key generation

updateKES

Arguments

∷ ∀ v m. (KESAlgorithm v, MonadST m, MonadThrow m)

⇒ ContextKES v

→ SignKeyKES v

→ Period
The current period for the key, not the target period.

→ m (Maybe (SignKeyKES v))

Update the KES signature key to the next period, given the current period.

It returns Nothing if the cannot be evolved any further.

The precondition (to get a Just result) is that the current KES period of the input key is not the last period. The given period must be the current KES period of the input key (not the next or target).

The postcondition is that in case a key is returned, its current KES period is incremented by one compared to before.

Note that you must track the current period separately, and to skip to a later period requires repeated use of this function, since it only increments one period at once.

forgetSignKeyKES ∷ (KESAlgorithm v, MonadST m, MonadThrow m) ⇒ SignKeyKES v → m ()

Forget a signing key synchronously, rather than waiting for GC. In some non-mock instances this provides a guarantee that the signing key is no longer in memory.

The precondition is that this key value will not be used again.

type Period = Word

The KES period. Periods are enumerated from zero.

Be careful of fencepost errors: if there are 2 periods (period 0 and 1) then there is only one key evolution.

class KESAlgorithm v ⇒ OptimizedKESAlgorithm v where

Subclass for KES algorithms that embed a copy of the VerKey into the signature itself, rather than relying on the externally supplied VerKey alone. Some optimizations made in the CompactSingleKES and CompactSumKES implementations require this additional interface in order to avoid redundant computations.

Methods

verifySigKES

Arguments

∷ (Signable v a, HasCallStack)

⇒ ContextKES v

→ Period
The current period for the key

→ a

→ SigKES v

→ Either String ()

Partial verification: this method only verifies the signature itself, but it does not check it against any externally-provided VerKey. Use verifyKES for full KES verification.

verKeyFromSigKES ∷ ContextKES v → Period → SigKES v → VerKeyKES v

Extract a VerKey from a SigKES. Note that a VerKey embedded in or derived from a SigKES is effectively user-supplied, so it is not enough to validate a SigKES against this VerKey (like verifySigKES does); you must also compare the VerKey against an externally-provided key that you want to verify against (see verifyKES).

verifyOptimizedKES ∷ (OptimizedKESAlgorithm v, Signable v a, HasCallStack) ⇒ ContextKES v → VerKeyKES v → Period → a → SigKES v → Either String ()

SignKeyWithPeriodKES wrapper

data SignKeyWithPeriodKES v

A sign key bundled with its associated period.

Constructors

SignKeyWithPeriodKES

Fields

skWithoutPeriodKES ∷ !(SignKeyKES v)

periodKES ∷ !Period

updateKESWithPeriod ∷ (KESAlgorithm v, MonadST m, MonadThrow m) ⇒ ContextKES v → SignKeyWithPeriodKES v → m (Maybe (SignKeyWithPeriodKES v))

SignedKES wrapper

newtype SignedKES v a

Constructors

SignedKES

Fields

getSig ∷ SigKES v

signedKES ∷ (KESAlgorithm v, Signable v a, MonadST m, MonadThrow m) ⇒ ContextKES v → Period → a → SignKeyKES v → m (SignedKES v a)

verifySignedKES ∷ (KESAlgorithm v, Signable v a) ⇒ ContextKES v → VerKeyKES v → Period → a → SignedKES v a → Either String ()

CBOR encoding and decoding

encodeVerKeyKES ∷ KESAlgorithm v ⇒ VerKeyKES v → Encoding

decodeVerKeyKES ∷ ∀ v s. KESAlgorithm v ⇒ Decoder s (VerKeyKES v)

encodeSigKES ∷ KESAlgorithm v ⇒ SigKES v → Encoding

decodeSigKES ∷ ∀ v s. KESAlgorithm v ⇒ Decoder s (SigKES v)

encodeSignedKES ∷ KESAlgorithm v ⇒ SignedKES v a → Encoding

decodeSignedKES ∷ KESAlgorithm v ⇒ Decoder s (SignedKES v a)

Encoded Size expressions

encodedVerKeyKESSizeExpr ∷ ∀ v. KESAlgorithm v ⇒ Proxy (VerKeyKES v) → Size

Size expression for VerKeyKES which is using sizeVerKeyKES encoded as Size.

encodedSignKeyKESSizeExpr ∷ ∀ v. KESAlgorithm v ⇒ Proxy (SignKeyKES v) → Size

Size expression for SignKeyKES which is using sizeSignKeyKES encoded as Size.

encodedSigKESSizeExpr ∷ ∀ v. KESAlgorithm v ⇒ Proxy (SigKES v) → Size

Size expression for SigKES which is using sizeSigKES encoded as Size.

Raw sizes

sizeVerKeyKES ∷ ∀ v proxy. KESAlgorithm v ⇒ proxy v → Word

sizeSigKES ∷ ∀ v proxy. KESAlgorithm v ⇒ proxy v → Word

sizeSignKeyKES ∷ ∀ v proxy. KESAlgorithm v ⇒ proxy v → Word

seedSizeKES ∷ ∀ v proxy. KESAlgorithm v ⇒ proxy v → Word

The upper bound on the Seed size needed by genKeyKES

Unsound APIs

class KESAlgorithm v ⇒ UnsoundKESAlgorithm v where

Unsound operations on KES sign keys. These operations violate secure forgetting constraints by leaking secrets to unprotected memory. Consider using the DirectSerialise / DirectDeserialise APIs instead.

Methods

rawDeserialiseSignKeyKESWith ∷ (MonadST m, MonadThrow m) ⇒ MLockedAllocator m → ByteString → m (Maybe (SignKeyKES v))

rawSerialiseSignKeyKES ∷ (MonadST m, MonadThrow m) ⇒ SignKeyKES v → m ByteString

encodeSignKeyKES ∷ ∀ v m. (UnsoundKESAlgorithm v, MonadST m, MonadThrow m) ⇒ SignKeyKES v → m Encoding

decodeSignKeyKES ∷ ∀ v s m. (UnsoundKESAlgorithm v, MonadST m, MonadThrow m) ⇒ Decoder s (m (Maybe (SignKeyKES v)))

rawDeserialiseSignKeyKES ∷ (UnsoundKESAlgorithm v, MonadST m, MonadThrow m) ⇒ ByteString → m (Maybe (SignKeyKES v))

class KESAlgorithm v ⇒ UnsoundPureKESAlgorithm v where

Pure implementations of the core KES operations. These are unsound, because proper handling of KES secrets (seeds, sign keys) requires mlocking and deterministic erasure ("secure forgetting"), which is not possible in pure code. This API is only provided for testing purposes; it must not be used to generate or use real KES keys.

Associated Types

data UnsoundPureSignKeyKES v ∷ Type

Methods

unsoundPureSignKES

Arguments

∷ ∀ a. Signable v a

⇒ ContextKES v

→ Period
The current period for the key

→ a

→ UnsoundPureSignKeyKES v

→ SigKES v

unsoundPureUpdateKES

Arguments

∷ ContextKES v

→ UnsoundPureSignKeyKES v

→ Period
The current period for the key, not the target period.

→ Maybe (UnsoundPureSignKeyKES v)

unsoundPureGenKeyKES ∷ Seed → UnsoundPureSignKeyKES v

unsoundPureDeriveVerKeyKES ∷ UnsoundPureSignKeyKES v → VerKeyKES v

unsoundPureSignKeyKESToSoundSignKeyKES ∷ (MonadST m, MonadThrow m) ⇒ UnsoundPureSignKeyKES v → m (SignKeyKES v)

rawSerialiseUnsoundPureSignKeyKES ∷ UnsoundPureSignKeyKES v → ByteString

rawDeserialiseUnsoundPureSignKeyKES ∷ ByteString → Maybe (UnsoundPureSignKeyKES v)

unsoundPureSignedKES ∷ (UnsoundPureKESAlgorithm v, Signable v a) ⇒ ContextKES v → Period → a → UnsoundPureSignKeyKES v → SignedKES v a

encodeUnsoundPureSignKeyKES ∷ UnsoundPureKESAlgorithm v ⇒ UnsoundPureSignKeyKES v → Encoding

decodeUnsoundPureSignKeyKES ∷ ∀ v s. UnsoundPureKESAlgorithm v ⇒ Decoder s (UnsoundPureSignKeyKES v)

Utility functions

hashPairOfVKeys ∷ (KESAlgorithm d, HashAlgorithm h) ⇒ (VerKeyKES d, VerKeyKES d) → Hash h (VerKeyKES d, VerKeyKES d)

mungeName ∷ String → String

unsoundPureSignKeyKESToSoundSignKeyKESViaSer ∷ (MonadST m, MonadThrow m, UnsoundKESAlgorithm k, UnsoundPureKESAlgorithm k) ⇒ UnsoundPureSignKeyKES k → m (SignKeyKES k)

Helper function for implementing unsoundPureSignKeyKESToSoundSignKeyKES for KES algorithms that support both UnsoundKESAlgorithm and UnsoundPureKESAlgorithm. For such KES algorithms, unsound sign keys can be marshalled to sound sign keys by serializing and then deserializing them.