Cardano.Crypto.KES.Class

Description

Abstract key evolving signatures.

Documentation

class (Typeable v, Show (VerKeyKES v), Eq (VerKeyKES v), Show (SignKeyKES v), Show (SigKES v), Eq (SigKES v), NoUnexpectedThunks (SigKES v), NoUnexpectedThunks (SignKeyKES v), NoUnexpectedThunks (VerKeyKES v)) ⇒ KESAlgorithm v where

Minimal complete definition

algorithmNameKES, deriveVerKeyKES, signKES, verifyKES, updateKES, currentPeriodKES, totalPeriodsKES, genKeyKES, seedSizeKES, (rawSerialiseVerKeyKES | encodeVerKeyKES), (rawSerialiseSignKeyKES | encodeSignKeyKES), (rawSerialiseSigKES | encodeSigKES), (rawDeserialiseVerKeyKES | decodeVerKeyKES), (rawDeserialiseSignKeyKES | decodeSignKeyKES), (rawDeserialiseSigKES | decodeSigKES)

Associated Types

data VerKeyKES v ∷ Type

data SignKeyKES v ∷ Type

data SigKES v ∷ Type

type ContextKES v ∷ Type

Context required to run the KES algorithm

Unit by default (no context required)

type Signable v ∷ Type → Constraint

Methods

algorithmNameKES ∷ proxy v → String

deriveVerKeyKES ∷ SignKeyKES v → VerKeyKES v

hashVerKeyKES ∷ HashAlgorithm h ⇒ VerKeyKES v → Hash h (VerKeyKES v)

signKES ∷ (Signable v a, HasCallStack) ⇒ ContextKES v → Natural → a → SignKeyKES v → Maybe (SigKES v)

verifyKES ∷ (Signable v a, HasCallStack) ⇒ ContextKES v → VerKeyKES v → Natural → a → SigKES v → Either String ()

updateKES ∷ HasCallStack ⇒ ContextKES v → SignKeyKES v → Natural → Maybe (SignKeyKES v)

Update the KES signature key to the specified period. The intended behavior is to return Nothing in the case that the key cannot be evolved that far.

The precondition is that the current KES period of the input key is before the target period. The postcondition is that in case a key is returned, its current KES period corresponds to the target KES period.

currentPeriodKES ∷ HasCallStack ⇒ ContextKES v → SignKeyKES v → Natural

Return the current KES period of a KES signing key.

totalPeriodsKES ∷ proxy v → Natural

Return the current KES period of a KES signing key.

genKeyKES ∷ Seed → SignKeyKES v

seedSizeKES ∷ proxy v → Natural

The upper bound on the Seed size needed by genKeyKES

rawSerialiseVerKeyKES ∷ VerKeyKES v → ByteString

rawSerialiseSignKeyKES ∷ SignKeyKES v → ByteString

rawSerialiseSigKES ∷ SigKES v → ByteString

rawDeserialiseVerKeyKES ∷ ByteString → Maybe (VerKeyKES v)

rawDeserialiseSignKeyKES ∷ ByteString → Maybe (SignKeyKES v)

rawDeserialiseSigKES ∷ ByteString → Maybe (SigKES v)

encodeVerKeyKES ∷ VerKeyKES v → Encoding

encodeSignKeyKES ∷ SignKeyKES v → Encoding

encodeSigKES ∷ SigKES v → Encoding

decodeVerKeyKES ∷ Decoder s (VerKeyKES v)

decodeSignKeyKES ∷ Decoder s (SignKeyKES v)

decodeSigKES ∷ Decoder s (SigKES v)

newtype SignedKES v a

Constructors

SignedKES

Fields

getSig ∷ SigKES v

signedKES ∷ (KESAlgorithm v, Signable v a) ⇒ ContextKES v → Natural → a → SignKeyKES v → Maybe (SignedKES v a)

verifySignedKES ∷ (KESAlgorithm v, Signable v a) ⇒ ContextKES v → VerKeyKES v → Natural → a → SignedKES v a → Either String ()

encodeSignedKES ∷ KESAlgorithm v ⇒ SignedKES v a → Encoding

decodeSignedKES ∷ KESAlgorithm v ⇒ Decoder s (SignedKES v a)