Cardano.Crypto.KES.Class

Description

Abstract key evolving signatures.

Documentation

class (Typeable v, Show (VerKeyKES v), Eq (VerKeyKES v), Show (SignKeyKES v), Show (SigKES v), Eq (SigKES v), NoUnexpectedThunks (SigKES v), NoUnexpectedThunks (SignKeyKES v), NoUnexpectedThunks (VerKeyKES v)) ⇒ KESAlgorithm v where

Associated Types

data VerKeyKES v ∷ Type

data SignKeyKES v ∷ Type

data SigKES v ∷ Type

type Signable v ∷ Type → Constraint

type ContextKES v ∷ Type

Context required to run the KES algorithm

Unit by default (no context required)

Methods

encodeVerKeyKES ∷ VerKeyKES v → Encoding

encodeSignKeyKES ∷ SignKeyKES v → Encoding

encodeSigKES ∷ SigKES v → Encoding

decodeVerKeyKES ∷ Decoder s (VerKeyKES v)

decodeSignKeyKES ∷ Decoder s (SignKeyKES v)

decodeSigKES ∷ Decoder s (SigKES v)

genKeyKES ∷ MonadRandom m ⇒ Natural → m (SignKeyKES v)

deriveVerKeyKES ∷ SignKeyKES v → VerKeyKES v

updateKES ∷ (MonadRandom m, HasCallStack) ⇒ ContextKES v → SignKeyKES v → Natural → m (Maybe (SignKeyKES v))

Update the KES signature key to the specified period. The intended behavior is to return Nothing in the case that the key cannot be evolved that far.

The precondition is that the current KES period of the input key is before the target period. The postcondition is that in case a key is returned, its current KES period corresponds to the target KES period.

signKES ∷ (MonadRandom m, Signable v a, HasCallStack) ⇒ ContextKES v → Natural → a → SignKeyKES v → m (Maybe (SigKES v))

verifyKES ∷ (Signable v a, HasCallStack) ⇒ ContextKES v → VerKeyKES v → Natural → a → SigKES v → Either String ()

currentPeriodKES ∷ HasCallStack ⇒ ContextKES v → SignKeyKES v → Natural

Return the current KES period of a KES signing key.

newtype SignedKES v a

Constructors

SignedKES

Fields

getSig ∷ SigKES v

signedKES ∷ (KESAlgorithm v, MonadRandom m, Signable v a) ⇒ ContextKES v → Natural → a → SignKeyKES v → m (Maybe (SignedKES v a))

verifySignedKES ∷ (KESAlgorithm v, Signable v a) ⇒ ContextKES v → VerKeyKES v → Natural → a → SignedKES v a → Either String ()

encodeSignedKES ∷ KESAlgorithm v ⇒ SignedKES v a → Encoding

decodeSignedKES ∷ KESAlgorithm v ⇒ Decoder s (SignedKES v a)