Cardano.Crypto.Libsodium

Initialization

sodiumInit :: IO ()

MLocked memory management

data MLockedForeignPtr a

Foreign pointer to securely allocated memory.

withMLockedForeignPtr :: forall a b. MLockedForeignPtr a -> (Ptr a -> IO b) -> IO b

allocMLockedForeignPtr :: Storable a => IO (MLockedForeignPtr a)

Allocate secure memory using c_sodium_malloc.

https://libsodium.gitbook.io/doc/memory_management

finalizeMLockedForeignPtr :: forall a. MLockedForeignPtr a -> IO ()

traceMLockedForeignPtr :: (Storable a, Show a) => MLockedForeignPtr a -> IO ()

Deprecated: Don't leave traceMLockedForeignPtr in production

MLocked bytes

data MLockedSizedBytes n

mlsbZero :: forall n. KnownNat n => MLockedSizedBytes n

Note: this doesn't need to allocate mlocked memory, but we do that for consistency

mlsbFromByteString :: forall n. KnownNat n => ByteString -> MLockedSizedBytes n

mlsbFromByteStringCheck :: forall n. KnownNat n => ByteString -> Maybe (MLockedSizedBytes n)

mlsbToByteString :: forall n. KnownNat n => MLockedSizedBytes n -> ByteString

Note: the resulting ByteString will still refer to secure memory, but the types don't prevent it from be exposed.

mlsbFinalize :: MLockedSizedBytes n -> IO ()

Calls finalizeMLockedForeignPtr on underlying pointer. This function invalidates argument.

Hashing

class HashAlgorithm h => SodiumHashAlgorithm h where

Methods

naclDigestPtr

Arguments

:: proxy h

-> Ptr a
input

-> Int
input length

-> IO (MLockedSizedBytes (SizeHash h))

digestMLockedStorable :: forall h a proxy. (SodiumHashAlgorithm h, Storable a) => proxy h -> Ptr a -> MLockedSizedBytes (SizeHash h)

digestMLockedBS :: forall h proxy. SodiumHashAlgorithm h => proxy h -> ByteString -> MLockedSizedBytes (SizeHash h)

expandHash :: forall h proxy. SodiumHashAlgorithm h => proxy h -> MLockedSizedBytes (SizeHash h) -> (MLockedSizedBytes (SizeHash h), MLockedSizedBytes (SizeHash h))