Cardano.Crypto.Libsodium.MLockedBytes.Internal

Documentation

newtype MLockedSizedBytes n

Constructors

MLSB (MLockedForeignPtr (PinnedSizedBytes n))

mlsbZero :: forall n. KnownNat n => MLockedSizedBytes n

Note: this doesn't need to allocate mlocked memory, but we do that for consistency

mlsbFromByteString :: forall n. KnownNat n => ByteString -> MLockedSizedBytes n

mlsbFromByteStringCheck :: forall n. KnownNat n => ByteString -> Maybe (MLockedSizedBytes n)

mlsbToByteString :: forall n. KnownNat n => MLockedSizedBytes n -> ByteString

Note: the resulting ByteString will still refer to secure memory, but the types don't prevent it from be exposed.

mlsbUseAsCPtr :: MLockedSizedBytes n -> (Ptr Word8 -> IO r) -> IO r

mlsbUseAsSizedPtr :: MLockedSizedBytes n -> (SizedPtr n -> IO r) -> IO r

mlsbFinalize :: MLockedSizedBytes n -> IO ()

Calls finalizeMLockedForeignPtr on underlying pointer. This function invalidates argument.