Cardano.Crypto.Seed

Description

Seeds for key generation.

Documentation

data Seed

A seed contains a finite number of bytes, and is used for seeding cryptographic algorithms including key generation.

This is not itself a PRNG, but can be used to seed a PRNG.

Constructing seeds

mkSeedFromBytes ∷ ByteString → Seed

Construct a Seed deterministically from a number of bytes.

getSeedBytes ∷ Seed → ByteString

readSeedFromSystemEntropy ∷ Word → IO Seed

Obtain a Seed by reading n bytes of entropy from the operating system.

splitSeed ∷ Word → Seed → Maybe (Seed, Seed)

Split a seed into two smaller seeds, the first of which is the given number of bytes large, and the second is the remaining. This will fail if not enough bytes are available. This can be chained multiple times provided the seed is big enough to cover each use.

expandSeed ∷ HashAlgorithm h ⇒ proxy h → Seed → (Seed, Seed)

Expand a seed into a pair of seeds using a cryptographic hash function (in the role of a crypto PRNG). The whole input seed is consumed. The output seeds are the size of the hash output.

Using seeds

getBytesFromSeed ∷ Word → Seed → Maybe (ByteString, Seed)

Get a number of bytes from the seed. This will fail if not enough bytes are available. This can be chained multiple times provided the seed is big enough to cover each use.

runMonadRandomWithSeed ∷ Seed → (∀ m. MonadRandom m ⇒ m a) → a

Run an action in MonadRandom deterministically using a seed as a finite source of randomness. Note that this is not a PRNG, so like with getBytesFromSeed it will fail if more bytes are requested than are available.

So this is only really suitable for key generation where there is a known upper bound on the amount of entropy that will be requested.

newtype SeedBytesExhausted

Constructors

SeedBytesExhausted

Fields

seedBytesSupplied ∷ Int